requirements.txt
playwright
openai
beautifulsoup4
lxml
pdfplumber
dnspython
python-dotenv
